Created login form
Created the form, added the code to check user credentials in the database (SQL query) and linked the form to the main application. Set the login form to appear on startup, before the main app. Only on successful login will you be able to access main app. If login unsuccessful, try again. If exit button is clicked the whole app stops and closes.